requirements.txt
asgiref==3.3.4
Django==3.2.4
gunicorn==20.1.0
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
django-widget-tweaks==1.4.8
daphne==3.0.2
channels==3.0.4
channels-redis==3.3.0
djangorestframework==3.12.4